Formatting change

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
 
-    public double calculate(double heightInMeters, double weightInKilograms) {
+    public int calculate(double heightInMeters, double weightInKilograms) {
         double index;
         index = weightInKilograms / (heightInMeters * heightInMeters);
